+++ src/main/kotlin/me/nobaboy/nobaaddons/NobaAddons.kt
@@ -1,6 +1,5 @@
 package me.nobaboy.nobaaddons
 
-import com.google.gson.Gson
 import com.mojang.logging.LogUtils
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineName
@@ -20,16 +19,16 @@ import me.nobaboy.nobaaddons.api.skyblock.SlayerAPI
 import me.nobaboy.nobaaddons.api.skyblock.events.mythological.BurrowAPI
 import me.nobaboy.nobaaddons.api.skyblock.events.mythological.BurrowGuessAPI
 import me.nobaboy.nobaaddons.api.skyblock.events.mythological.DianaAPI
+import me.nobaboy.nobaaddons.api.skyblock.fishing.SeaCreatureAPI
 import me.nobaboy.nobaaddons.api.skyblock.fishing.TrophyFishAPI
 import me.nobaboy.nobaaddons.commands.NobaCommand
 import me.nobaboy.nobaaddons.commands.SWikiCommand
 import me.nobaboy.nobaaddons.config.NobaConfig
-import me.nobaboy.nobaaddons.config.util.safeLoad
 import me.nobaboy.nobaaddons.config.UISettings
+import me.nobaboy.nobaaddons.config.util.safeLoad
 import me.nobaboy.nobaaddons.core.PersistentCache
 import me.nobaboy.nobaaddons.core.UpdateNotifier
 import me.nobaboy.nobaaddons.features.chat.CopyChatFeature
-import me.nobaboy.nobaaddons.features.chat.alerts.IAlert
 import me.nobaboy.nobaaddons.features.chat.chatcommands.impl.DMCommands
 import me.nobaboy.nobaaddons.features.chat.chatcommands.impl.GuildCommands
 import me.nobaboy.nobaaddons.features.chat.chatcommands.impl.PartyCommands
@@ -43,12 +42,14 @@ import me.nobaboy.nobaaddons.features.events.hoppity.HoppityEggGuess
 import me.nobaboy.nobaaddons.features.events.mythological.AnnounceRareDrops
 import me.nobaboy.nobaaddons.features.events.mythological.BurrowWaypoints
 import me.nobaboy.nobaaddons.features.events.mythological.InquisitorWaypoints
+import me.nobaboy.nobaaddons.features.fishing.AnnounceSeaCreatures
 import me.nobaboy.nobaaddons.features.fishing.CatchTimer
-import me.nobaboy.nobaaddons.features.fishing.CatchMessageModifications
 import me.nobaboy.nobaaddons.features.fishing.FishingBobberTweaks
-import me.nobaboy.nobaaddons.features.fishing.HighlightThunderSparks
+import me.nobaboy.nobaaddons.features.fishing.HotspotWaypoints
+import me.nobaboy.nobaaddons.features.fishing.RevertTreasureMessages
 import me.nobaboy.nobaaddons.features.fishing.SeaCreatureAlert
-import me.nobaboy.nobaaddons.features.fishing.TrophyFishChat
+import me.nobaboy.nobaaddons.features.fishing.crimsonisle.HighlightThunderSparks
+import me.nobaboy.nobaaddons.features.fishing.crimsonisle.TrophyFishChat
 import me.nobaboy.nobaaddons.features.inventory.ItemPickupLog
 import me.nobaboy.nobaaddons.features.inventory.enchants.EnchantmentTooltips
 import me.nobaboy.nobaaddons.features.inventory.slotinfo.ISlotInfo
@@ -66,7 +67,7 @@ import me.nobaboy.nobaaddons.features.slayers.inferno.HighlightHellionShield
 import me.nobaboy.nobaaddons.features.slayers.sven.HidePupNametags
 import me.nobaboy.nobaaddons.features.slayers.voidgloom.VoidgloomSeraphFeatures
 import me.nobaboy.nobaaddons.features.ui.infobox.InfoBoxesManager
-import me.nobaboy.nobaaddons.features.visuals.EtherwarpHelper
+import me.nobaboy.nobaaddons.features.visuals.EtherwarpOverlay
 import me.nobaboy.nobaaddons.features.visuals.TemporaryWaypoints
 import me.nobaboy.nobaaddons.repo.RepoManager
 import me.nobaboy.nobaaddons.ui.UIManager
@@ -99,8 +100,6 @@ object NobaAddons : ClientModInitializer {
 	val LOGGER: Logger = LogUtils.getLogger()
 	val CONFIG_DIR: Path get() = FabricLoader.getInstance().configDir.resolve(MOD_ID)
 
-	val GSON = Gson()
-
 	@OptIn(ExperimentalSerializationApi::class)
 	val JSON = Json {
 		ignoreUnknownKeys = true
@@ -129,7 +128,7 @@ object NobaAddons : ClientModInitializer {
 	// immediately ran).
 	override fun onInitializeClient() {
 		/* region Core */
-		NobaConfig.INSTANCE.safeLoad()
+		NobaConfig.safeLoad()
 		PersistentCache.safeLoad()
 		RepoManager.init()
 		UISettings.safeLoad()
@@ -148,6 +147,7 @@ object NobaAddons : ClientModInitializer {
 		MayorAPI.init()
 		PartyAPI.init()
 		PetAPI.init()
+		SeaCreatureAPI.init()
 		SkyBlockAPI.init()
 		SlayerAPI.init()
 		TrophyFishAPI.init()
@@ -170,7 +170,7 @@ object NobaAddons : ClientModInitializer {
 
 		/* region Features */
 		// region Visuals
-		EtherwarpHelper.init()
+		EtherwarpOverlay.init()
 		TemporaryWaypoints.init()
 		// endregion
 
@@ -208,10 +208,12 @@ object NobaAddons : ClientModInitializer {
 		// endregion
 
 		// region Fishing
+		AnnounceSeaCreatures.init()
 		CatchTimer.init()
-		CatchMessageModifications.init()
 		FishingBobberTweaks.init()
 		HighlightThunderSparks.init()
+		HotspotWaypoints.init()
+		RevertTreasureMessages.init()
 		SeaCreatureAlert.init()
 		TrophyFishChat.init()
 		// endregion
@@ -230,7 +232,6 @@ object NobaAddons : ClientModInitializer {
 		// region Chat
 		CopyChatFeature.init()
 		ChatNotifications.init()
-		IAlert.init()
 		IChatFilter.init()
 		/* region Chat Commands */
 		DMCommands.init()