+++ src/main/kotlin/me/nobaboy/nobaaddons/config/categories/FishingCategory.kt
@@ -7,10 +7,11 @@ import me.nobaboy.nobaaddons.config.util.*
 import me.nobaboy.nobaaddons.core.Rarity
 import me.nobaboy.nobaaddons.core.Rarity.Companion.toArray
 import me.nobaboy.nobaaddons.core.fishing.TrophyFishRarity
-import me.nobaboy.nobaaddons.features.fishing.TrophyFishChat
+import me.nobaboy.nobaaddons.features.fishing.crimsonisle.TrophyFishChat
 import me.nobaboy.nobaaddons.ui.TextShadow
 import me.nobaboy.nobaaddons.utils.CommonText
 import me.nobaboy.nobaaddons.utils.TextUtils.white
+import me.nobaboy.nobaaddons.utils.enums.AnnounceChannel
 import me.nobaboy.nobaaddons.utils.tr
 import net.minecraft.text.Text
 
@@ -22,27 +23,77 @@ object FishingCategory {
 			booleanController()
 		}
 
+		add({ fishing::hotspotWaypoints }) {
+			name = tr("nobaaddons.config.fishing.hotspotWaypoints", "Hotspot Waypoints")
+			descriptionText = tr("nobaaddons.config.fishing.hotspotWaypoints.tooltip", "Marks hotspots with a beacon and shows a timer when it's going to disappear")
+			booleanController()
+		}
+
 		add({ fishing.catchMessages::revertTreasureMessages }) {
 			name = tr("nobaaddons.config.fishing.catchMessages.revertTreasureMessages", "Revert Treasure Catch Messages")
 			descriptionText = tr("nobaaddons.config.fishing.catchMessages.revertTreasureMessages.tooltip", "Reverts treasure catch messages to the format used pre-Backwater Bayou")
 			booleanController()
 		}
 
+		seaCreatureAlert()
+		announceSeaCreatures()
 		bobberTimer()
 		catchTimer()
 		trophyFishing()
-		seaCreatureAlert()
 		highlightThunderSparks()
+	}
+
+	private fun ConfigCategory.Builder.seaCreatureAlert() {
+		group(tr("nobaaddons.config.fishing.seaCreatureAlert", "Sea Creature Alert")) {
+			val enabled = add({ fishing.seaCreatureAlert::enabled }) {
+				name = CommonText.Config.ENABLED
+				booleanController()
+			}
+			add({ fishing.seaCreatureAlert::nameInsteadOfRarity }) {
+				name = tr("nobaaddons.config.fishing.seaCreatureAlert.nameInsteadOfRarity", "Use Name Instead of Rarity")
+				descriptionText = tr("nobaaddons.config.fishing.seaCreatureAlert.nameInsteadOfRarity.tooltip", "Uses the sea creature's name instead when displaying the notification, instead of 'Legendary Catch!'")
+				require { option(enabled) }
+				booleanController()
+			}
+			add({ fishing.seaCreatureAlert::minimumRarity }) {
+				name = tr("nobaaddons.config.fishing.seaCreatureAlert.minimumRarity", "Minimum Rarity")
+				descriptionText = tr("nobaaddons.config.fishing.seaCreatureAlert.minimumRarity.tooltip", "The minimum rarity to display a catch notification for")
+				require { option(enabled) }
+				enumController(onlyInclude = (Rarity.COMMON..Rarity.MYTHIC).toArray())
+			}
+			add({ fishing.seaCreatureAlert::notificationSound }) {
+				name = CommonText.Config.NOTIFICATION_SOUND
+				require { option(enabled) }
+				enumController()
+			}
+		}
+	}
+
+	private fun ConfigCategory.Builder.announceSeaCreatures() {
+		group(tr("nobaaddons.config.fishing.announceSeaCreatures", "Announce Sea Creatures")) {
+			val enabled = add({ fishing.announceSeaCreatures::enabled }) {
+				name = CommonText.Config.ENABLED
+				booleanController()
+			}
+			add({ fishing.announceSeaCreatures::minimumRarity }) {
+				name = tr("nobaaddons.config.fishing.announceSeaCreatures.minimumRarity", "Minimum Rarity")
+				descriptionText = tr("nobaaddons.config.fishing.announceSeaCreatures.minimumRarity.tooltip", "The minimum rarity to announce for")
+				require { option(enabled) }
+				enumController(onlyInclude = (Rarity.LEGENDARY..Rarity.MYTHIC).toArray())
+			}
+			add({ fishing.announceSeaCreatures::announceChannel }) {
+				name = CommonText.Config.ANNOUNCE_CHANNEL
+				require { option(enabled) }
+				enumController(onlyInclude = arrayOf(AnnounceChannel.ALL, AnnounceChannel.PARTY))
+			}
+		}
 	}
 
 	private fun ConfigCategory.Builder.bobberTimer() {
 		group(tr("nobaaddons.config.fishing.bobberTimer", "Bobber Timer")) {
 			val enabled = add({ fishing.bobberTimer::enabled }) {
 				name = CommonText.Config.ENABLED
-				descriptionText = tr(
-					"nobaaddons.config.fishing.bobberTimer.tooltip",
-					"Displays how long your fishing bobber has been spawned above it; the color will change to gold once it's been spawned long enough to catch a Slugfish."
-				)
+				descriptionText = tr("nobaaddons.config.fishing.bobberTimer.tooltip", "Displays how long your fishing bobber has been spawned above it; the color will change to gold once it's been spawned long enough to catch a Slugfish.")
 				booleanController()
 			}
 			add({ fishing.bobberTimer::crimsonIsleOnly }) {
@@ -87,44 +138,6 @@ object FishingCategory {
 				name = tr("nobaaddons.config.fishing.trophyFishing.compactMaxRarity", "Compact Max Rarity")
 				descriptionText = tr("nobaaddons.config.fishing.trophyFishing.compactMaxRarity.tooltip", "The maximum rarity to compact catch messages for. If this is set to Diamond, this will effectively compact all catch messages.")
 				require { option(modify) and option(compact) }
-				enumController()
-			}
-		}
-	}
-
-	private fun ConfigCategory.Builder.seaCreatureAlert() {
-		group(tr("nobaaddons.config.fishing.seaCreatureAlert", "Sea Creature Alert")) {
-			val enabled = add({ fishing.seaCreatureAlert::enabled }) {
-				name = CommonText.Config.ENABLED
-				booleanController()
-			}
-			add({ fishing.seaCreatureAlert::nameInsteadOfRarity }) {
-				name = tr("nobaaddons.config.fishing.seaCreatureAlert.nameInsteadOfRarity", "Use Name instead of Rarity")
-				descriptionText = tr("nobaaddons.config.fishing.seaCreatureAlert.nameInsteadOfRarity.tooltip", "Uses the sea creature's name instead when displaying the notification, instead of 'Legendary Catch!'")
-				require { option(enabled) }
-				booleanController()
-			}
-			add({ fishing.seaCreatureAlert::minimumRarity }) {
-				name = tr("nobaaddons.config.fishing.seaCreatureAlert.minimumRarity", "Minimum Rarity")
-				descriptionText = tr("nobaaddons.config.fishing.seaCreatureAlert.minimumRarity.tooltip", "The minimum rarity to display a catch notification for")
-				require { option(enabled) }
-				enumController(onlyInclude = (Rarity.COMMON..Rarity.MYTHIC).toArray())
-			}
-			add({ fishing.seaCreatureAlert::carrotKingIsRare }) {
-				name = tr("nobaaddons.config.fishing.seaCreatureAlert.carrotKingIsRare", "Carrot King is Rare")
-				descriptionText = tr("nobaaddons.config.fishing.seaCreatureAlert.carrotKingIsRare.tooltip", "Carrot King will be considered rare even if the above minimum rarity isn't low enough for it (since not many people fish for it)")
-				require { option(enabled) }
-				booleanController()
-			}
-			add({ fishing.seaCreatureAlert::announceInPartyChat }) {
-				name = tr("nobaaddons.config.fishing.seaCreatureAlert.announceInPartyChat", "Announce in Party Chat")
-				descriptionText = tr("nobaaddons.config.fishing.seaCreatureAlert.announceInPartyChat.tooltip", "A chat message will also be sent in party chat when catching a rare creature")
-				require { option(enabled) }
-				booleanController()
-			}
-			add({ fishing.seaCreatureAlert::notificationSound }) {
-				name = CommonText.Config.NOTIFICATION_SOUND
-				require { option(enabled) }
 				enumController()
 			}
 		}