+++ src/main/kotlin/me/nobaboy/nobaaddons/config/categories/InventoryCategory.kt
@@ -2,8 +2,8 @@ package me.nobaboy.nobaaddons.config.categories
 
 import dev.isxander.yacl3.api.Binding
 import dev.isxander.yacl3.api.ConfigCategory
-import me.nobaboy.nobaaddons.config.util.*
 import me.nobaboy.nobaaddons.config.UISettings
+import me.nobaboy.nobaaddons.config.util.*
 import me.nobaboy.nobaaddons.features.inventory.enchants.EnchantmentDisplayMode
 import me.nobaboy.nobaaddons.ui.TextShadow
 import me.nobaboy.nobaaddons.utils.CommonText
@@ -12,7 +12,7 @@ import me.nobaboy.nobaaddons.utils.tr
 object InventoryCategory {
 	fun create() = category(tr("nobaaddons.config.inventory", "Inventory")) {
 		slotInfo()
-		enchantTooltips()
+		enchantmentTooltips()
 		itemPickupLog()
 	}
 
@@ -135,7 +135,7 @@ object InventoryCategory {
 		}
 	}
 
-	private fun ConfigCategory.Builder.enchantTooltips() {
+	private fun ConfigCategory.Builder.enchantmentTooltips() {
 		group(tr("nobaaddons.config.inventory.enchantmentTooltips", "Enchantment Tooltips")) {
 			val enabled = add({ inventory.enchantmentTooltips::modifyTooltips }) {
 				name = tr("nobaaddons.config.inventory.enchantmentTooltips.modifyTooltips", "Modify Enchant Tooltips")