+++ src/main/kotlin/me/nobaboy/nobaaddons/config/configs/EventsConfig.kt
@@ -1,36 +1,35 @@
 package me.nobaboy.nobaaddons.config.configs
 
-import dev.celestialfault.celestialconfig.ObjectProperty
-import dev.celestialfault.celestialconfig.Property
-import dev.celestialfault.celestialconfig.Serializer
+import dev.celestialfault.histoire.Object
+import me.nobaboy.nobaaddons.utils.enums.AnnounceChannel
 import me.nobaboy.nobaaddons.utils.sound.NotificationSound
 
-class EventsConfig : ObjectProperty<EventsConfig>("events") {
-	val hoppity by Hoppity()
-	val mythological by Mythological()
+class EventsConfig {
+	@Object val hoppity = Hoppity()
+	@Object val mythological = Mythological()
 
-	class Hoppity : ObjectProperty<Hoppity>("hoppity") {
-		var eggGuess by Property.of<Boolean>("eggGuess", false)
-		var requireMythicRabbit by Property.of<Boolean>("requireMythicRabbit", false)
+	class Hoppity {
+		var eggGuess = false
+		var requireMythicRabbit = false
 	}
 
-	class Mythological : ObjectProperty<Mythological>("mythological") {
-		var burrowGuess by Property.of<Boolean>("burrowGuess", false)
-		var findNearbyBurrows by Property.of<Boolean>("findNearbyBurrows", false)
-		var dingOnBurrowFind by Property.of<Boolean>("dingOnBurrowFind", false)
-		var removeGuessOnBurrowFind by Property.of<Boolean>("removeGuessOnBurrowFind", false)
-		var findNearestWarp by Property.of<Boolean>("findNearestWarp", false)
+	class Mythological {
+		var burrowGuess = false
+		var findNearbyBurrows = false
+		var dingOnBurrowFind = false
+		var removeGuessOnBurrowFind = false
+		var findNearestWarp = false
 
-		var alertInquisitor by Property.of<Boolean>("alertInquisitor", false)
-		var alertOnlyInParty by Property.of<Boolean>("alertOnlyInParty", false)
-		var notificationSound by Property.of("notificationSound", Serializer.enum(), NotificationSound.DING)
-		var showInquisitorDespawnTime by Property.of<Boolean>("showInquisitorDespawnTime", false)
-		var inquisitorFocusMode by Property.of<Boolean>("inquisitorFocusMode", false)
+		var alertInquisitor = false
+		var announceChannel = AnnounceChannel.PARTY
+		var notificationSound = NotificationSound.DING
+		var showInquisitorDespawnTime = false
+		var inquisitorFocusMode = false
 
-		var announceRareDrops by Property.of<Boolean>("announceRareDrops", false)
+		var announceRareDrops = false
 
-		var ignoreCrypt by Property.of<Boolean>("ignoreCrypt", false)
-		var ignoreWizard by Property.of<Boolean>("ignoreWizard", false)
-		var ignoreStonks by Property.of<Boolean>("ignoreStonks", false)
+		var ignoreCrypt = false
+		var ignoreWizard = false
+		var ignoreStonks = false
 	}
 }