+++ src/main/kotlin/me/nobaboy/nobaaddons/config/configs/QOLConfig.kt
@@ -1,29 +1,28 @@
 package me.nobaboy.nobaaddons.config.configs
 
-import dev.celestialfault.celestialconfig.ObjectProperty
-import dev.celestialfault.celestialconfig.Property
+import dev.celestialfault.histoire.Object
 
-class QOLConfig : ObjectProperty<QOLConfig>("qol") {
-	val soundFilters by SoundFilters()
-	val garden by Garden()
+class QOLConfig {
+	@Object val soundFilters = SoundFilters()
+	@Object val garden = Garden()
 
-	class SoundFilters : ObjectProperty<SoundFilters>("soundFilters") {
-		var muteWitherSkullAbilities by Property.of<Boolean>("muteWitherSkullAbilities", false)
+	class SoundFilters {
+		var muteWitherSkullAbilities = false
 
-		var muteBanshee by Property.of<Boolean>("muteBanshee", false)
-		var muteReindrakeSpawn by Property.of<Boolean>("muteReindrakeSpawn", false)
-		var muteReindrakeGiftDrop by Property.of<Boolean>("muteReindrakeGiftDrop", false)
+		var muteBanshee = false
+		var muteReindrakeSpawn = false
+		var muteReindrakeGiftDrop = false
 
-		var muteGoneWithTheWind by Property.of<Boolean>("muteGoneWithTheWind", false)
+		var muteGoneWithTheWind = false
 
-		var muteKillerSpring by Property.of<Boolean>("muteKillerSpring", false)
+		var muteKillerSpring = false
 
-		var mutePunch by Property.of<Boolean>("mutePunch", false)
+		var mutePunch = false
 	}
 
-	class Garden : ObjectProperty<Garden>("garden") {
-		var reduceMouseSensitivity by Property.of<Boolean>("reduceMouseSensitivity", false)
-		var reductionMultiplier by Property.of<Int>("reductionMultiplier", 6)
-		var autoUnlockMouseOnTeleport by Property.of<Boolean>("autoUnlockMouseOnTeleport", false)
+	class Garden {
+		var reduceMouseSensitivity = false
+		var reductionMultiplier = 6
+		var autoUnlockMouseOnTeleport = false
 	}
 }