+++ src/main/kotlin/me/nobaboy/nobaaddons/config/util/io.kt
@@ -1,8 +1,7 @@
 package me.nobaboy.nobaaddons.config.util
 
-import dev.celestialfault.celestialconfig.AbstractConfig
+import dev.celestialfault.histoire.Histoire
 import me.nobaboy.nobaaddons.NobaAddons
-import me.nobaboy.nobaaddons.mixins.accessors.AbstractConfigAccessor
 import me.nobaboy.nobaaddons.utils.ErrorManager
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents
 import net.minecraft.util.PathUtil
@@ -12,39 +11,40 @@ import java.time.format.DateTimeFormatter
 import java.util.Locale
 import kotlin.io.path.nameWithoutExtension
 
-private val DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ROOT)
+@PublishedApi
+internal val DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ROOT)
 
 /**
- * Attempts to load the associated [AbstractConfig], logging an error and renaming the config file if it fails.
+ * Attempts to call [loader], logging an error and renaming [path] if an error is encountered.
  */
-fun AbstractConfig.safeLoad(pathSupplier: () -> Path = { (this as AbstractConfigAccessor).callGetPath() }) {
-	try {
-		load()
-	} catch(ex: Throwable) {
-		val path = pathSupplier()
-		ErrorManager.logError("Failed to load a config file", ex)
+inline fun <R> safeLoad(path: Path, loader: () -> R): Result<R> = runCatching(loader).onFailure {
+	ErrorManager.logError("Failed to load a config file", it)
 
-		val date = DATE_FORMATTER.format(ZonedDateTime.now())
-		val name = "${path.nameWithoutExtension}-${date}"
-		val backup = PathUtil.getNextUniqueName(NobaAddons.CONFIG_DIR, name, ".json")
+	val date = DATE_FORMATTER.format(ZonedDateTime.now())
+	val name = "${path.nameWithoutExtension}-${date}"
+	val backup = PathUtil.getNextUniqueName(NobaAddons.CONFIG_DIR, name, ".json")
 
-		if(path.toFile().renameTo(path.parent.resolve(backup).toFile())) {
-			NobaAddons.LOGGER.warn("Moved config file to $backup")
-		} else {
-			NobaAddons.LOGGER.warn("Couldn't rename config file")
-		}
+	if(path.toFile().renameTo(path.parent.resolve(backup).toFile())) {
+		NobaAddons.LOGGER.warn("Moved config file to $backup")
+	} else {
+		NobaAddons.LOGGER.warn("Couldn't rename config file")
 	}
 }
 
 /**
- * Attaches a [ClientLifecycleEvents] listener for when the client is stopping which calls [AbstractConfig.save]
+ * Attempts to load the associated [Histoire] instance, logging an error and renaming the file if it fails.
  */
-fun AbstractConfig.saveOnExit(onlyIfDirty: Boolean = false) {
+fun Histoire.safeLoad() = safeLoad(file.toPath(), ::load)
+
+/**
+ * Attaches a [ClientLifecycleEvents] listener for when the client is stopping which calls [Histoire.save]
+ */
+fun Histoire.saveOnExit() {
 	ClientLifecycleEvents.CLIENT_STOPPING.register {
-		if(onlyIfDirty && !dirty) return@register
 		try {
 			save()
 		} catch(ex: Throwable) {
+			// nothing we can do at this point other than just log an error and give up
 			NobaAddons.LOGGER.error("Failed to save ${this::class.simpleName} before shutdown", ex)
 		}
 	}