+++ src/main/kotlin/me/nobaboy/nobaaddons/core/UpdateNotifier.kt
@@ -26,7 +26,7 @@ import me.nobaboy.nobaaddons.utils.tr
 import net.fabricmc.loader.api.Version
 
 object UpdateNotifier {
-	private val config get() = NobaConfig.INSTANCE.general
+	private val config get() = NobaConfig.general
 
 	private var inWorld: Boolean = false
 	private var notifiedUpdate: Boolean = false
@@ -59,7 +59,8 @@ object UpdateNotifier {
 	fun sendUpdateNotification() {
 		val update = UPDATE_INFO ?: return
 		if(MCUtils.VERSION_INFO !in update.forMinecraft) return
-		ChatUtils.addMessage(buildText {
+
+		val message = buildText {
 			appendLine()
 			append(NobaAddons.PREFIX)
 			append(tr("nobaaddons.updateAvailable", "A new update is available: ${update.latest}:").gold())
@@ -70,7 +71,9 @@ object UpdateNotifier {
 				.openUrl("https://modrinth.com/mod/nobaaddons/versions")
 				.underline())
 			appendLine()
-		}, prefix = false)
+		}
+
+		ChatUtils.addMessage(message, prefix = false)
 	}
 
 	@Serializable