+++ src/main/kotlin/me/nobaboy/nobaaddons/features/chat/CopyChatFeature.kt
@@ -10,13 +10,14 @@ import me.nobaboy.nobaaddons.utils.chat.ChatUtils
 import me.nobaboy.nobaaddons.utils.tr
 import net.minecraft.client.gui.hud.ChatHudLine
 import net.minecraft.client.gui.screen.Screen
+import net.minecraft.text.Text
 import java.lang.ref.WeakReference
 import java.util.WeakHashMap
 
 object CopyChatFeature {
-	private val messages = WeakHashMap<ChatHudLine.Visible, WeakReference<ChatHudLine>>(200)
+	private val config get() = NobaConfig.chat.copyChat
 
-	private val config get() = NobaConfig.INSTANCE.chat.copyChat
+	private val messages = WeakHashMap<ChatHudLine.Visible, WeakReference<ChatHudLine>>(200)
 
 	fun init() {
 		ChatMessageEvents.ADDED.register(this::onChatAdded)
@@ -54,6 +55,7 @@ object CopyChatFeature {
 		} else {
 			ChatUtils.addMessage(tr("nobaaddons.chat.copiedMessage", "Copied chat message to clipboard"))
 		}
+
 		return true
 	}
 
@@ -62,7 +64,7 @@ object CopyChatFeature {
 		CTRL_CLICK,
 		;
 
-		override fun getDisplayName() = when(this) {
+		override fun getDisplayName(): Text = when(this) {
 			RIGHT_CLICK -> tr("nobaaddons.label.copyButton.rightClick", "Right Click")
 			CTRL_CLICK -> tr("nobaaddons.label.copyButton.ctrlClick", "Control Held")
 		}