+++ src/main/kotlin/me/nobaboy/nobaaddons/features/dungeons/HighlightStarredMobs.kt
@@ -16,7 +16,7 @@ import net.minecraft.entity.decoration.ArmorStandEntity
 
 // TODO: Rework and implement Entity outlines
 object HighlightStarredMobs {
-	private val config get() = NobaConfig.INSTANCE.dungeons.highlightStarredMobs
+	private val config get() = NobaConfig.dungeons.highlightStarredMobs
 	private val enabled: Boolean get() = config.enabled && !DungeonsAPI.inBoss()
 
 	private val starredMobs = mutableListOf<ArmorStandEntity>()
@@ -33,7 +33,7 @@ object HighlightStarredMobs {
 		EntityUtils.getEntities<ArmorStandEntity>().filter {
 			it !in starredMobs &&
 			it.hasCustomName() &&
-			it.shouldRenderName() == true &&
+			it.shouldRenderName() &&
 			it.customName!!.string.cleanFormatting().startsWith("✯ ") &&
 			it.customName!!.string.cleanFormatting().endsWith("❤")
 		}.forEach { starredMobs.add(it) }