+++ src/main/kotlin/me/nobaboy/nobaaddons/features/fishing/RevertTreasureMessages.kt
@@ -1,5 +1,6 @@
 package me.nobaboy.nobaaddons.features.fishing
 
+import me.nobaboy.nobaaddons.api.skyblock.SkyBlockAPI
 import me.nobaboy.nobaaddons.config.NobaConfig
 import me.nobaboy.nobaaddons.events.impl.chat.ChatMessageEvents
 import me.nobaboy.nobaaddons.repo.Repo.fromRepo
@@ -8,18 +9,19 @@ import me.nobaboy.nobaaddons.utils.TextUtils.buildText
 import net.minecraft.text.MutableText
 import net.minecraft.util.Formatting
 
-object CatchMessageModifications {
-	private val config by NobaConfig.INSTANCE.fishing::catchMessages
+object RevertTreasureMessages {
+	private val config get() = NobaConfig.fishing.catchMessages
+	private val enabled: Boolean get() = config.revertTreasureMessages && !SkyBlockAPI.inSkyBlock
 
-	private val TREASURE_CATCH_MESSAGE by Regex("^⛃ (?<rarity>GOOD|GREAT|OUTSTANDING)(?<treasureType> JUNK)? CATCH! You caught .+").fromRepo("fishing.treasure_catch")
+	private val TREASURE_CATCH_REGEX by Regex("^⛃ (?<rarity>GOOD|GREAT|OUTSTANDING)(?<treasureType> JUNK)? CATCH! You caught .+").fromRepo("fishing.treasure_catch")
 
-	private val colorOverride = mapOf(
+	private val treasureCatchColors = mapOf(
 		"GOOD" to Formatting.GOLD,
 		"GREAT" to Formatting.DARK_PURPLE,
 	)
 
 	fun init() {
-		ChatMessageEvents.MODIFY.register(::modifyMessage)
+		ChatMessageEvents.MODIFY.register(this::modifyMessage)
 	}
 
 	/*
@@ -45,21 +47,20 @@ object CatchMessageModifications {
 	 */
 
 	private fun modifyMessage(event: ChatMessageEvents.Modify) {
-		if(!config.revertTreasureMessages) return
+		if(!enabled) return
 
 		val message = event.message
-		val match = TREASURE_CATCH_MESSAGE.matchEntire(message.string.cleanFormatting()) ?: return
+		val match = TREASURE_CATCH_REGEX.matchEntire(message.string.cleanFormatting()) ?: return
 		val siblings = message.siblings.toMutableList()
 		siblings.removeFirst() // drop the icon
 
 		val rarity = match.groups["rarity"]!!.value
-		val overrideColor = colorOverride[rarity].also(::println) ?: return
+		val catchColor = treasureCatchColors[rarity] ?: return
 		val isJunk = match.groups["treasureType"] != null
 
-		(siblings[0] as MutableText).styled { it.withColor(overrideColor) }
-		if(isJunk) {
-			(siblings[2] as MutableText).styled { it.withColor(overrideColor) }
-		}
+		// revert the color of the catch tier
+		(siblings[0] as MutableText).styled { it.withColor(catchColor) }
+		if(isJunk) (siblings[2] as MutableText).styled { it.withColor(catchColor) }
 
 		event.message = buildText {
 			siblings.forEach(::append)