+++ src/main/kotlin/me/nobaboy/nobaaddons/features/mining/glacitemineshaft/MineshaftWaypointType.kt
@@ -3,5 +3,6 @@ package me.nobaboy.nobaaddons.features.mining.glacitemineshaft
 enum class MineshaftWaypointType {
 	ENTRANCE,
 	LADDER,
-	CORPSE
+	CORPSE,
+	;
 }