+++ src/main/kotlin/me/nobaboy/nobaaddons/features/mining/glacitemineshaft/MineshaftWaypoints.kt
@@ -15,7 +15,7 @@ import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderEvents
 import net.minecraft.text.Text
 
 object MineshaftWaypoints {
-	private val config get() = NobaConfig.INSTANCE.mining.glaciteMineshaft
+	private val config get() = NobaConfig.mining.glaciteMineshaft
 
 	val waypoints = mutableListOf<MineshaftWaypoint>()
 
@@ -53,12 +53,11 @@ object MineshaftWaypoints {
 				MineshaftWaypointType.ENTRANCE -> config.entranceWaypoint
 				MineshaftWaypointType.LADDER -> config.ladderWaypoint
 				MineshaftWaypointType.CORPSE -> config.corpseLocator
-				else -> false
 			}
 
 			if(!shouldRender) return
 			RenderUtils.renderWaypoint(context, it.location, it.color, throughBlocks = true)
-			RenderUtils.renderText(it.location.center().raise(), it.text, yOffset = -5f, hideThreshold = 5.0, throughBlocks = true)
+			RenderUtils.renderText(context, it.location.center().raise(), it.text, yOffset = -5f, hideThreshold = 5.0, throughBlocks = true)
 		}
 	}