+++ src/main/kotlin/me/nobaboy/nobaaddons/features/ui/infobox/InfoBoxElement.kt
@@ -1,22 +1,15 @@
 package me.nobaboy.nobaaddons.features.ui.infobox
 
-import com.google.gson.JsonObject
-import dev.celestialfault.celestialconfig.Property
-import me.nobaboy.nobaaddons.ui.data.AbstractTextElement
+import kotlinx.serialization.Serializable
+import me.nobaboy.nobaaddons.ui.TextShadow
 import me.nobaboy.nobaaddons.ui.data.ElementPosition
+import me.nobaboy.nobaaddons.ui.data.TextElement
 
-class InfoBoxElement() : AbstractTextElement<InfoBoxElement>() {
-	constructor(position: ElementPosition) : this() {
-		this.position.x = position.x
-		this.position.y = position.y
-		this.position.scale = position.scale
-	}
-
-	constructor(json: JsonObject) : this() {
-		load(json)
-	}
-
-	var text by Property.of("text", "")
-
-	fun copy() = InfoBoxElement(save() as JsonObject)
-}
+@Serializable
+data class InfoBoxElement(
+	override var textShadow: TextShadow = TextShadow.SHADOW,
+	override var color: Int = 0xFFFFFF,
+	override var outlineColor: Int = 0x000000,
+	override val position: ElementPosition = ElementPosition(),
+	var text: String = "",
+) : TextElement