+++ src/main/kotlin/me/nobaboy/nobaaddons/features/visuals/EtherwarpOverlay.kt
@@ -1,5 +1,6 @@
 package me.nobaboy.nobaaddons.features.visuals
 
+import dev.isxander.yacl3.api.NameableEnum
 import me.nobaboy.nobaaddons.api.skyblock.SkyBlockAPI
 import me.nobaboy.nobaaddons.config.NobaConfig
 import me.nobaboy.nobaaddons.utils.LocationUtils.rayCast
@@ -19,8 +20,8 @@ import net.minecraft.util.hit.BlockHitResult
 import net.minecraft.util.hit.HitResult
 import net.minecraft.world.World
 
-object EtherwarpHelper {
-	private val config get() = NobaConfig.INSTANCE.uiAndVisuals.etherwarpHelper
+object EtherwarpOverlay {
+	private val config get() = NobaConfig.uiAndVisuals.etherwarpOverlay
 	private val enabled: Boolean get() = config.enabled && SkyBlockAPI.inSkyBlock && MCUtils.options.sneakKey.isPressed
 
 	private const val BASE_DISTANCE = 57
@@ -73,7 +74,7 @@ object EtherwarpHelper {
 		targetBlock = validateTargetBlock(world, target)
 		if(targetBlock == ValidationType.TOO_FAR && !config.allowOverlayOnAir) return
 
-		val color = targetBlock?.let { NobaColor.GRAY } ?: config.highlightColor
+		val color = targetBlock?.let { config.failHighlightColor } ?: config.highlightColor
 		RenderUtils.renderOutlinedFilledBox(context, target.blockPos.toNobaVec(), color, throughBlocks = true)
 	}
 
@@ -90,9 +91,16 @@ object EtherwarpHelper {
 		return if(stateAbove.isSolid && !stateAbove.isAir || !stateTwoAbove.isAir) ValidationType.NO_AIR_ABOVE else null
 	}
 
-	private enum class ValidationType(val displayName: Text) {
-		TOO_FAR(tr("nobaaddons.config.uiAndVisuals.etherwarpHelper.validationType.tooFar", "Too far!")),
-		NOT_SOLID(tr("nobaaddons.config.uiAndVisuals.etherwarpHelper.validationType.notSolid", "Not solid!")),
-		NO_AIR_ABOVE(tr("nobaaddons.config.uiAndVisuals.etherwarpHelper.validationType.noAirAbove", "No air above!"))
+	private enum class ValidationType : NameableEnum {
+		TOO_FAR,
+		NOT_SOLID,
+		NO_AIR_ABOVE,
+		;
+
+		override fun getDisplayName(): Text = when(this) {
+			TOO_FAR -> tr("nobaaddons.config.uiAndVisuals.etherwarpOverlay.validationType.tooFar", "Too far!")
+			NOT_SOLID -> tr("nobaaddons.config.uiAndVisuals.etherwarpOverlay.validationType.notSolid", "Not solid!")
+			NO_AIR_ABOVE -> tr("nobaaddons.config.uiAndVisuals.etherwarpOverlay.validationType.noAirAbove", "No air above!")
+		}
 	}
 }